merged_bot/biblia/requirements.txt
python-telegram-bot[job-queue]==21.9
python-dotenv==1.0.0
schedule==1.2.0
pytz==2024.1
